refactor: CarFactory 패키지 이동
