+++ src/main/java/racingcar/service/RacingCarService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 import racingcar.dao.CarDao;
 import racingcar.dao.GameDao;
 import racingcar.domain.Car;
+import racingcar.domain.CarFactory;
 import racingcar.domain.Cars;
 import racingcar.domain.RacingGame;
 import racingcar.dto.CarDto;
